WIP: Integration tests
