+++ AddressesAPI.Tests/IntegrationTests.cs
@@ -1,4 +1,4 @@
-using System;
+    using System;
 using System.IO;
 using System.Net.Http;
 using AddressesAPI.V1.Infrastructure;
