--- AddressesAPI.Tests/V1/E2ETests/GetAddressCrossReferenceE2ETests.cs
@@ -1,35 +1,0 @@
-using AddressesAPI.V1.Boundary.Responses;
-using FluentAssertions;
-using Newtonsoft.Json;
-using NUnit.Framework;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Threading.Tasks;
-
-namespace AddressesAPI.Tests.V1.E2ETests
-{
-    public class GetAddressCrossReferenceE2ETests : IntegrationTests<Startup>
-    {
-        [Ignore("in progress")]
-        [Test]
-        public async Task GetCrossReferenceAddressReturns200()
-        {
-            var response = await CallEndpointWithQueryParameters().ConfigureAwait(true);
-            response.StatusCode.Should().Be(200);
-        }
-
-        private async Task<HttpResponseMessage> CallEndpointWithQueryParameters()
-        {
-            var url = new Uri("api/v1/properties/{uprn}/crossreferences", UriKind.Relative);
-            return await Client.GetAsync(url).ConfigureAwait(true);
-        }
-
-        private static async Task<GetAddressCrossReferenceResponse> ConvertToResponseObject(HttpResponseMessage response)
-        {
-            var data = await response.Content.ReadAsStringAsync().ConfigureAwait(true);
-            return JsonConvert.DeserializeObject<GetAddressCrossReferenceResponse>(data);
-        }
-    }
-}
