+++ AddressesAPI/V1/Domain/NotFoundException.cs
@@ -1,0 +1,11 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace AddressesAPI.V1.Domain
+{
+    public class NotFoundException : Exception
+    {
+    }
+}
